bringing in compliance with the deploy branch

+++ backend/accounts/views.py
@@ -1,5 +1,7 @@
 from rest_framework.views import APIView
 from rest_framework import status, permissions
+from rest_framework.authentication import TokenAuthentication
+from rest_framework.permissions import AllowAny
 from rest_framework.authtoken.models import Token
 from django.contrib.auth import authenticate
 from drugs.utils.custom_response import CustomResponse
@@ -28,8 +30,22 @@ class LoginUser(APIView):
             )
 
         token, created = Token.objects.get_or_create(user=user)
+
+        if user.is_superuser:
+            role = "superuser"
+ 
+        elif user.is_staff:
+            role = "staff"
+ 
+        else:
+            role = "no_role"
+
         return CustomResponse.response(
-            data={"token": token.key},
+            data={
+                "token": token.key, 
+                "username": username, 
+                "role": role,
+            },
             status=200,
             message="Авторизация прошла успешно",
             http_status=status.HTTP_200_OK
@@ -37,13 +53,17 @@ class LoginUser(APIView):
 
 
 class LogoutUser(APIView):
+    permission_classes = [AllowAny]
     permission_classes = [permissions.IsAuthenticated]
 
     def post(self, request):
-        try:
-            request.user.auth_token.delete()
-        except Token.DoesNotExist:
-            pass  
+        user = request.user
+
+        if user and user.is_authenticated:
+            try:
+                request.user.auth_token.delete()
+            except Token.DoesNotExist:
+                pass  
 
         return CustomResponse.response(
             message="Выход выполнен успешно",
